test: Extract throttle object to separate header
